feat(button): Add numberOfLines and ellipsizeMode to Text in Buttons
I've added the possibility to use the two following options in a button:
- textNumberOfLines
- textEllipsizeMode
My own use case is that I render buttons based on a posts title. If a
title is too long, I get a multiline button, which breaks the layout.
This way, it's possible to keep one at it one line.